+++ src/StateSmithTest/DrawIoSvgToSmDiagramConverterTests.cs
@@ -4,6 +4,8 @@ using StateSmith.Input.DrawIo;
 using System.Linq;
 using FluentAssertions;
 using StateSmith.Runner;
+using StateSmith.Compiling;
+using StateSmith.compiler;
 
 namespace StateSmithTest.DrawIo;
 
@@ -30,6 +32,7 @@ public class DrawIoSvgToSmDiagramConverterTests
         runner.CompileNodesToVertices(converter.Roots, converter.Edges);
         runner.FindSingleStateMachine();
         runner.sm.Name.Should().Be("Design1Sm_svg");
+        ValidateSm(runner);
     }
 
     [Fact]
@@ -42,6 +45,7 @@ public class DrawIoSvgToSmDiagramConverterTests
         runner.FindSingleStateMachine();
 
         runner.sm.Name.Should().Be("Design1Sm_svg");
+        ValidateSm(runner);
     }
 
     [Fact]
@@ -56,6 +60,7 @@ public class DrawIoSvgToSmDiagramConverterTests
         runner.FindSingleStateMachine();
 
         runner.sm.Name.Should().Be("Design1Sm_compressed");
+        ValidateSm(runner);
     }
 
     [Fact]
@@ -70,6 +75,7 @@ public class DrawIoSvgToSmDiagramConverterTests
         runner.FindSingleStateMachine();
 
         runner.sm.Name.Should().Be("Design1Sm_regular");
+        ValidateSm(runner);
     }
 
     [Fact]
@@ -84,6 +90,71 @@ public class DrawIoSvgToSmDiagramConverterTests
         runner.FindSingleStateMachine();
 
         runner.sm.Name.Should().Be("Design1Sm_regular_dio");
+        ValidateSm(runner);
+    }
+
+    [Fact]
+    public void CompileDrawIoFileNodesToVertices_WithImage()
+    {
+        string filePath = ExamplesTestHelpers.TestInputDirectoryPath + "drawio/Design1Sm_with_image.drawio.svg";
+
+        CompilerRunner runner = new();
+        runner.CompileDrawIoFileNodesToVertices(filePath);
+        runner.FindSingleStateMachine();
+        runner.sm.Name.Should().Be("Design1Sm_with_image");
+
+        runner.RemoveNotesVertices();
+        ValidateSm(runner);
+    }
+
+    private static void ValidateSm(CompilerRunner runner)
+    {
+        var sm = runner.sm;
+
+        State GetState(string stateName)
+        {
+            return (State)sm.Descendant(stateName);
+        }
+
+        sm.ShouldHaveChildrenAndUmlBehaviors("ROOT.InitialState, OFF, ON_GROUP, BOOM", "");
+
+        GetState("OFF").ShouldHaveChildrenAndUmlBehaviors("", $$"""
+            enter / { light_off(); }
+            INCREASE TransitionTo(ON1)
+            """
+        );
+
+        GetState("ON_GROUP").ShouldHaveChildrenAndUmlBehaviors("ON1, ON2, ON3", $$"""
+            OFF TransitionTo(OFF)
+            """
+        );
+
+        GetState("ON1").ShouldHaveChildrenAndUmlBehaviors("", $$"""
+            enter / { light_on1(); }
+            INCREASE TransitionTo(ON2)
+            DIM TransitionTo(OFF)
+            """
+        );
+
+        GetState("ON2").ShouldHaveChildrenAndUmlBehaviors("", $$"""
+            enter / { light_on2(); }
+            INCREASE TransitionTo(ON3)
+            DIM TransitionTo(ON1)
+            """
+        );
+
+        GetState("ON3").ShouldHaveChildrenAndUmlBehaviors("", $$"""
+            1. INCREASE / { count++; }
+            2. INCREASE [count >= 3] TransitionTo(BOOM)
+            enter / { count = 0;\nlight_on3(); }
+            DIM TransitionTo(ON2)
+            """
+        );
+
+        GetState("BOOM").ShouldHaveChildrenAndUmlBehaviors("", $$"""
+            enter / { light_boom(); }
+            """
+        );
     }
 }