+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [WIP-alpha]
+### Added
+- draw.io - embedded images inside of a draw.io diagram are now ignored.
+  - images are sometimes embedded in state machine diagrams for documentation purposes.
+
+---
+
 ## [0.7.1-alpha]
 ### Added
 - Add initial support for draw.io files as alternative to yEd files. https://github.com/StateSmith/StateSmith/issues/77